% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{catch_em_app}
\alias{catch_em_app}
\title{Run \code{catch_em()} with \code{shiny}}
\usage{
catch_em_app(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
A \code{shiny} app object.
}
\description{
Run \code{catch_em()} interactively.
}
\author{
Almog Simchon
}
