% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_em.R
\name{compare_txt}
\alias{compare_txt}
\title{Match cheaters}
\usage{
compare_txt(txt1, txt2, n_grams = 10, across = c("both", "txt1", "txt2"))
}
\arguments{
\item{txt1, txt2}{character vectors to compare, each of length 1.}

\item{n_grams}{see \link{ngram} package.}

\item{across}{How should the percentage of overlap be computed?}
}
\value{
The percent (0-1) of overlap between the texts
}
\description{
Match cheaters
}
\examples{
text1 <- "My horse is large and white, and I ride it every day."
text2 <- "My mule is large and brown, and I ride it most days."
compare_txt(text1, text2, n_grams = 3)

}
\author{
Mattan S. Ben-Shachar
}
