% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{graph_em}
\alias{graph_em}
\title{plot cheatrs}
\usage{
graph_em(object, weight_range = c(0.7, 1), ...)
}
\arguments{
\item{object}{output of \code{catch_em()}.}

\item{weight_range}{range of edge values to plot}

\item{...}{not used}
}
\description{
plot cheatrs
}
\author{
Mattan S. Ben-Shachar
}
