% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hist.chtrs}
\alias{hist.chtrs}
\title{plot histogram of similarity scores}
\usage{
\method{hist}{chtrs}(object, ...)
}
\arguments{
\item{object}{output of \code{catch_em()}.}

\item{...}{passed to \code{hist}}
}
\description{
plot histogram of similarity scores
}
\author{
Mattan S. Ben-Shachar
}
