% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.chtrs}
\alias{print.chtrs}
\title{Print Cheatrs}
\usage{
print.chtrs(x, ...)
}
\arguments{
\item{x}{output of \code{catch_em()}.}

\item{...}{not used}
}
\description{
Print Cheatrs
}
\author{
Mattan S. Ben-Shachar
}
