% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chtrs}
\alias{summary.chtrs}
\title{Summarise Cheatrs}
\usage{
\method{summary}{chtrs}(object, bad_files = FALSE)
}
\arguments{
\item{bad_files}{logical. Instead of the result matrix, should return instead the list of bad files (that did not compare / load)? default \code{FALSE}}

\item{x}{output of \code{catch_em()}.}
}
\description{
Summarise Cheatrs
}
\author{
Mattan S. Ben-Shachar
}
