% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_em.R
\name{catch_em}
\alias{catch_em}
\title{Match cheaters}
\usage{
catch_em(flist, n_grams = 10, time_lim = 1L)
}
\arguments{
\item{...}{}
}
\value{
matrix
}
\description{
Match cheaters
}
\author{
Mattan S. Ben-Shachar
}
