% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_em.R
\name{compare_txt}
\alias{compare_txt}
\title{Match cheaters}
\usage{
compare_txt(txt1, txt2, n_grams = 10)
}
\arguments{
\item{txt1}{a char vector of length 1 to compare to `txt2`}

\item{txt2}{a char vector of length 1 to compare to `txt1`}
}
\value{
The percent (0-1) of overlap between the texts
}
\description{
Match cheaters
}
\author{
Mattan S. Ben-Shachar
}
