% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.chtrs}
\alias{plot.chtrs}
\alias{graph_em}
\alias{hist.chtrs}
\title{Plot cheatrs / histogram of similarity scores}
\usage{
\method{plot}{chtrs}(x, weight_range = c(0.4, 1), remove_lonely = TRUE, digits = 0, ...)

graph_em(x, weight_range = c(0.4, 1), remove_lonely = TRUE, digits = 0, ...)

\method{hist}{chtrs}(x, ...)
}
\arguments{
\item{x}{output of \code{\link[=catch_em]{catch_em()}}.}

\item{weight_range}{range of edge values to plot}

\item{remove_lonely}{should lonely nodes (with no edges) be removed from the
graph?}

\item{digits}{Number of digits to round the percentage to.}

\item{...}{passed to \code{\link[ggraph:ggraph]{ggraph::ggraph()}}, \link[ggplot2:geom_histogram]{ggplot2::geom_histogram} or
\code{\link[=hist]{hist()}}.}
}
\description{
\code{plot} requires \code{ggraph} and \code{ggplot2} to work.
}
\author{
Mattan S. Ben-Shachar
}
