% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chtrs}
\alias{summary.chtrs}
\title{print cheatrs}
\usage{
\method{summary}{chtrs}(x, bad_files = F)
}
\arguments{
\item{x}{output of \code{catch_em()}.}

\item{bad_files}{logical. Instead of the result matrix, should return instead the list of bad files (that did not compare / load)? default \code{FALSE}}
}
\description{
print cheatrs
}
\author{
Mattan S. Ben-Shachar
}
