% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chtrs}
\alias{summary.chtrs}
\title{Summarise Cheatrs}
\usage{
\method{summary}{chtrs}(object, bad_files = FALSE, ...)
}
\arguments{
\item{object}{output of \code{\link[=catch_em]{catch_em()}}.}

\item{bad_files}{logical. Instead of the result matrix, should return instead
the list of bad files (that did not compare / load)? Defaults to \code{FALSE}.}

\item{...}{Not used.}
}
\value{
The input \code{chtrs} matrix, or a list of bad files (when \code{bad_files = TRUE}).
}
\description{
Summarise Cheatrs
}
\examples{

if (interactive()) {
  files <- choose.files()
  res <- catch_em(files)

  summary(res, bad_files = TRUE)
}

}
\author{
Mattan S. Ben-Shachar
}
