% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chtrs}
\alias{summary.chtrs}
\title{Summarise Cheatrs}
\usage{
\method{summary}{chtrs}(object, bad_files = FALSE, ...)
}
\arguments{
\item{object}{output of \code{\link[=catch_em]{catch_em()}}.}

\item{bad_files}{logical. Instead of the result matrix, should return instead
the list of bad files (that did not compare / load)? Defaults to \code{FALSE}.}

\item{...}{Not used.}
}
\description{
Summarise Cheatrs
}
\author{
Mattan S. Ben-Shachar
}
